Remove duplicated logics and move the function to check end.
