+++ src/main/scala/com/databricks/spark/xml/parsers/StaxXmlParserUtils.scala
@@ -1,7 +1,9 @@
 package com.databricks.spark.xml.parsers
 
 import javax.xml.stream.XMLEventReader
-import javax.xml.stream.events.XMLEvent
+import javax.xml.stream.events.{Characters, StartElement, EndElement, XMLEvent}
+
+import com.databricks.spark.xml.XmlOptions
 
 private[xml] object StaxXmlParserUtils {
   /**
@@ -28,4 +30,28 @@ private[xml] object StaxXmlParserUtils {
     }
     data
   }
+
+  /**
+   * Check if current event points the EndElement.
+   */
+  def checkEndElement(parser: XMLEventReader, options: XmlOptions): Boolean = {
+    val current = parser.peek
+    current match {
+      case _: EndElement => true
+      case _: StartElement => false
+      case _: Characters =>
+        // When `Characters` is found here, we need to look further to decide
+        // if this is really `EndElement` because this can be whitespace between
+        // `EndElement` and `StartElement`.
+        val next = {
+          parser.nextEvent
+          parser.peek
+        }
+        next match {
+          case _: EndElement => true
+          case _: XMLEvent => false
+        }
+    }
+  }
+
 }
